chore(@clayui/date-picker): update DatePicker component stories
